mes-3652-extract shared components and refactor some existing components
