+++ src/pages/waiting-room-to-car/cat-b/components/instructor-registration/instructor-registration.html
