+++ src/pages/waiting-room-to-car/cat-b/components/tell-me-question-outcome/tell-me-question-outcome.ts
