+++ src/pages/waiting-room-to-car/cat-b/components/tell-me-question/tell-me-question.ts
@@ -1,6 +1,6 @@
 import { Component, Input, Output, EventEmitter, OnChanges } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
-import { TellMeQuestion } from '../../../../providers/question/tell-me-question.model';
+import { TellMeQuestion } from '../../../../../providers/question/tell-me-question.model';
 
 @Component({
   selector: 'tell-me-question',
@@ -24,7 +24,7 @@ export class TellMeQuestionComponent implements OnChanges {
 
   ngOnChanges(): void {
     if (!this.formControl) {
-      this.formControl  = new FormControl('TellMeQuestion', [Validators.required]);
+      this.formControl = new FormControl('TellMeQuestion', [Validators.required]);
       this.formGroup.addControl('tellMeQuestion', this.formControl);
     }
     this.formControl.patchValue(this.tellMeQuestion);
