+++ src/pages/waiting-room-to-car/cat-b/components/transmission/transmission.ts
