+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.cat-b.page.ts
@@ -322,4 +322,8 @@ export class WaitingRoomToCarCatBPage extends PracticeableBasePageComponent {
     this.store$.dispatch(action);
   }
 
+  getDebriefPage() {
+    return CAT_B.DEBRIEF_PAGE;
+  }
+
 }
